# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{projectId}/prompts/{promptId}/versions/{versionTag}/translations/{languageCode}": {
      "get": {
        "operationId": "PromptTranslationController_findOneByLanguage",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "Prompt CUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code (e.g., es-ES)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptTranslationDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific translation by language code for a prompt version",
        "tags": [
          "Prompt Translations (within Project/Prompt/Version)"
        ]
      },
      "patch": {
        "operationId": "PromptTranslationController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "Prompt CUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code of the translation to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptTranslationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptTranslationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a specific translation by language code for a prompt version",
        "tags": [
          "Prompt Translations (within Project/Prompt/Version)"
        ]
      },
      "delete": {
        "operationId": "PromptTranslationController_remove",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "Prompt CUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code of the translation to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation deleted."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a specific translation by language code for a prompt version",
        "tags": [
          "Prompt Translations (within Project/Prompt/Version)"
        ]
      }
    },
...
    "/api/projects/{projectId}/assets/{assetKey}/versions/{versionTag}/translations": {
      "post": {
        "operationId": "AssetTranslationController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Asset Key",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetTranslationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Translation created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAssetTranslationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Asset, or Version not found."
          },
          "409": {
            "description": "Translation for this language already exists for this version."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a translation for a specific asset version within a project",
        "tags": [
          "Asset Translations (within Project/Asset/Version)"
        ]
      },
      "get": {
        "operationId": "AssetTranslationController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Asset Key",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of translations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateAssetTranslationDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Asset, or Version not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all translations for a specific asset version within a project",
        "tags": [
          "Asset Translations (within Project/Asset/Version)"
        ]
      }
    },
    "/api/projects/{projectId}/assets/{assetKey}/versions/{versionTag}/translations/{languageCode}": {
      "get": {
        "operationId": "AssetTranslationController_findOneByLanguage",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Asset Key",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code (e.g., es-ES)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAssetTranslationDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Asset, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific translation by language code for an asset version",
        "tags": [
          "Asset Translations (within Project/Asset/Version)"
        ]
      },
      "patch": {
        "operationId": "AssetTranslationController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Asset Key",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code of the translation to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssetTranslationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAssetTranslationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Asset, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a specific translation by language code for an asset version",
        "tags": [
          "Asset Translations (within Project/Asset/Version)"
        ]
      },
      "delete": {
        "operationId": "AssetTranslationController_remove",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Asset Key",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code of the translation to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation deleted."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Asset, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a specific translation by language code for an asset version",
        "tags": [
          "Asset Translations (within Project/Asset/Version)"
        ]
      }
    },
...
  },
  "info": {
    "title": "JAPM API",
    "description": "API para la aplicación de Prompt Engineering JAPM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreatePromptTranslationDto": {
        "type": "object",
        "properties": {
          "versionId": {
            "type": "string",
            "description": "ID de la versión del prompt a la que pertenece esta traducción",
            "example": "cl...........cuid"
          },
          "languageCode": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "Código de idioma para esta traducción (formato xx-XX)",
            "example": "es-ES"
          },
          "promptText": {
            "type": "string",
            "description": "Texto del prompt traducido a este idioma"
          }
        },
        "required": [
          "versionId",
          "languageCode",
          "promptText"
        ]
      },
      "UpdatePromptTranslationDto": {
        "type": "object",
        "properties": {
          "promptText": {
            "type": "string",
            "description": "Texto del prompt traducido a este idioma"
          }
        }
      },
...
      "CreateAssetTranslationDto": {
        "type": "object",
        "properties": {
          "versionId": {
            "type": "string",
            "description": "ID de la versión del asset a la que pertenece esta traducción",
            "example": "cl...........cuid"
          },
          "languageCode": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "Código de idioma para esta traducción (formato xx-XX)",
            "example": "es-ES"
          },
          "value": {
            "type": "string",
            "description": "Valor del asset traducido a este idioma"
          }
        },
        "required": [
          "versionId",
          "languageCode",
          "value"
        ]
      },
      "UpdateAssetTranslationDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Valor del asset traducido a este idioma"
          }
        }
      },
...
    }
  }
}